requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.3
asn1crypto==1.2.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
conda==4.8.0
conda-package-handling==1.6.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.1
django-ckeditor==5.8.0
django-controlcenter==0.2.9
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
django-material-admin==1.6.23
django-pkgconf==0.3.0
django-restful-admin==1.0.3
django-webpush==0.3.1
djangorestframework==3.11.0
graphviz==0.13.2
gunicorn==20.0.4
hiredis==1.0.1
http-ece==1.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1.1
menuinst==1.4.16
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
Pillow==6.2.1
py-vapid==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.19
pygraphviz==1.5
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
pywebpush==1.9.4
pywin32==223
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
ruamel-yaml==0.15.46
ruamel.yaml==0.15.46
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
style==1.1.0
tqdm==4.36.1
Twisted==19.10.0
txaio==18.8.1
update==0.0.1
urllib3==1.24.2
whitenoise==5.0.1
win-inet-pton==1.1.0
wincertstore==0.2
zope.interface==4.7.1
